Change everything to usize

+++ src/day1.rs
@@ -3,7 +3,7 @@ use crate::{Solution, Test};
 pub struct Day1;
 
 impl Solution for Day1 {
-    type Input = Vec<Vec<u32>>;
+    type Input = Vec<Vec<usize>>;
 
     fn parse(input: &str) -> Self::Input {
         input
@@ -15,20 +15,20 @@ impl Solution for Day1 {
             .collect()
     }
 
-    fn part1(input: &Self::Input) -> u32 {
+    fn part1(input: &Self::Input) -> usize {
         input.iter().map(|elf| elf.iter().sum()).max().unwrap()
     }
 
-    fn part2(input: &Self::Input) -> u32 {
-        let mut input: Vec<u32> = input.iter().map(|elf| elf.iter().sum()).collect();
+    fn part2(input: &Self::Input) -> usize {
+        let mut input: Vec<usize> = input.iter().map(|elf| elf.iter().sum()).collect();
         input.sort_by(|a, b| a.cmp(b).reverse());
         input.iter().take(3).sum()
     }
 }
 
 impl Test for Day1 {
-    const TEST_OUTPUT1: u32 = 24000;
-    const TEST_OUTPUT2: u32 = 45000;
+    const TEST_OUTPUT1: usize = 24000;
+    const TEST_OUTPUT2: usize = 45000;
     const TEST_INPUT: &'static str = "
         1000
         2000
